requirements.txt
streamlit
streamlit-ace
google-generativeai
google-genai
python-dotenv